comlink fails in Angular, works for React, Svelte and Vue
This repository contains two worker packages: one that is a regular
worker and another that has the worker wrapped by comlink. The packages
are prebuild.
Running the several test apps for React, Angular, Svelte and Vue it can
be seen that the regular worker is working everywhere (it is important
to disable dependency optimization for the packages in vite bundled apps
though). The comlink wrapped worker is running everywhere except in
Angular.